fork master

+++ tcc-transaction-core/src/main/java/org/mengyun/tcctransaction/recover/TransactionRecovery.java
@@ -1,12 +1,15 @@
 package org.mengyun.tcctransaction.recover;
 
+import com.alibaba.fastjson.JSON;
+import org.apache.commons.lang3.exception.ExceptionUtils;
 import org.apache.log4j.Logger;
+import org.mengyun.tcctransaction.OptimisticLockException;
 import org.mengyun.tcctransaction.Transaction;
 import org.mengyun.tcctransaction.TransactionRepository;
 import org.mengyun.tcctransaction.api.TransactionStatus;
+import org.mengyun.tcctransaction.common.TransactionType;
 import org.mengyun.tcctransaction.support.TransactionConfigurator;
 
-import java.util.ArrayList;
 import java.util.Calendar;
 import java.util.Date;
 import java.util.List;
@@ -39,18 +42,7 @@ public class TransactionRecovery {
 
         List<Transaction> transactions = transactionRepository.findAllUnmodifiedSince(new Date(currentTimeInMillis - transactionConfigurator.getRecoverConfig().getRecoverDuration() * 1000));
 
-        List<Transaction> recoverTransactions = new ArrayList<Transaction>();
-
-        for (Transaction transaction : transactions) {
-
-            int result = transactionRepository.update(transaction);
-
-            if (result > 0) {
-                recoverTransactions.add(transaction);
-            }
-        }
-
-        return recoverTransactions;
+        return transactions;
     }
 
     private void recoverErrorTransactions(List<Transaction> transactions) {
@@ -60,27 +52,46 @@ public class TransactionRecovery {
 
             if (transaction.getRetriedCount() > transactionConfigurator.getRecoverConfig().getMaxRetryCount()) {
 
-                logger.error(String.format("recover failed with max retry count,will not try again. txid:%s, status:%s,retried count:%d", transaction.getXid(), transaction.getStatus().getId(), transaction.getRetriedCount()));
+                logger.error(String.format("recover failed with max retry count,will not try again. txid:%s, status:%s,retried count:%d,transaction content:%s", transaction.getXid(), transaction.getStatus().getId(), transaction.getRetriedCount(), JSON.toJSONString(transaction)));
                 continue;
             }
 
+            if (transaction.getTransactionType().equals(TransactionType.BRANCH)
+                    && (transaction.getCreateTime().getTime() +
+                    transactionConfigurator.getRecoverConfig().getMaxRetryCount() *
+                            transactionConfigurator.getRecoverConfig().getRecoverDuration() * 1000
+                    > System.currentTimeMillis())) {
+                continue;
+            }
+
+            
             try {
                 transaction.addRetriedCount();
 
                 if (transaction.getStatus().equals(TransactionStatus.CONFIRMING)) {
+
                     transaction.changeStatus(TransactionStatus.CONFIRMING);
                     transactionConfigurator.getTransactionRepository().update(transaction);
                     transaction.commit();
+                    transactionConfigurator.getTransactionRepository().delete(transaction);
 
-                } else {
+                } else if (transaction.getStatus().equals(TransactionStatus.CANCELLING)
+                        || transaction.getTransactionType().equals(TransactionType.ROOT)) {
+                    
                     transaction.changeStatus(TransactionStatus.CANCELLING);
                     transactionConfigurator.getTransactionRepository().update(transaction);
                     transaction.rollback();
+                    transactionConfigurator.getTransactionRepository().delete(transaction);
                 }
 
-                transactionConfigurator.getTransactionRepository().delete(transaction);
-            } catch (Throwable e) {
-                logger.warn(String.format("recover failed, txid:%s, status:%s,retried count:%d", transaction.getXid(), transaction.getStatus().getId(), transaction.getRetriedCount()), e);
+            } catch (Throwable throwable) {
+
+                if (throwable instanceof OptimisticLockException
+                        || ExceptionUtils.getRootCause(throwable) instanceof OptimisticLockException) {
+                    logger.warn(String.format("optimisticLockException happened while recover. txid:%s, status:%s,retried count:%d,transaction content:%s", transaction.getXid(), transaction.getStatus().getId(), transaction.getRetriedCount(), JSON.toJSONString(transaction)), throwable);
+                } else {
+                    logger.error(String.format("recover failed, txid:%s, status:%s,retried count:%d,transaction content:%s", transaction.getXid(), transaction.getStatus().getId(), transaction.getRetriedCount(), JSON.toJSONString(transaction)), throwable);
+                }
             }
         }
     }
